refactor: organizing code structure
